model upgrade + http patch working + protected against under posting

+++ movtrack8-backend/Models/TEpisode.cs
@@ -12,7 +12,7 @@ namespace movtrack8_backend.Models
         public TWebsite Website { get; set; }
         public long WebsiteId { get; set; }
 
-        public int JackettId { get; set; }
+        public long JackettId { get; set; }
 
         public DateTime PubDate { get; set; }
         public string Title { get; set; }
